Time: 100 ms (12.69%), Space: 79.2 MB (19.8%) - LeetHub

+++ 2626-count-the-number-of-good-subarrays/2626-count-the-number-of-good-subarrays.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long countGood(vector<int>& nums, int k) {
         long long ans =0;
-        long long count=0;
+        int count=0;
         int left=0;
         int right = 0;
         unordered_map<int,int> m;
